+++ supertokens_python/recipe/session/session_functions.py
@@ -19,7 +19,6 @@ from typing import TYPE_CHECKING, Any, Dict, List, Union, Optional
 from supertokens_python.recipe.session.interfaces import SessionInformationResult
 
 from .access_token import get_info_from_access_token
-from .constants import JWKCacheMaxAgeInMs
 from .jwt import ParsedJWTInfo
 
 if TYPE_CHECKING:
@@ -159,6 +158,7 @@ async def get_session(
 
     try:
         access_token_info = get_info_from_access_token(
+            config,
             parsed_access_token,
             config.anti_csrf_function_or_string == "VIA_TOKEN" and do_anti_csrf_check,
         )
@@ -198,7 +198,8 @@ async def get_session(
 
             # We check if the token was created since the last time we refreshed the keys from the core
             # Since we do not know the exact timing of the last refresh, we check against the max age
-            if time_created <= time.time() - JWKCacheMaxAgeInMs:
+
+            if time_created <= time.time() - config.default_jwk_refresh_interval_sec:
                 raise e
         else:
             # Since v3 (and above) tokens contain a kid we can trust the cache refresh mechanism built on top of the pyjwt lib